pkg: move k8s dependent command in a subpackage
Since the `podres` pkg pulls in ~32M of deps, bringing
the total executable size to a shopping 40M, push the offending command
in a subpacakge and link it from the client code.

This way, code willing to use the core knit functionality doesn't have
to pay the full price anyway
